remove old nxls client and fix showing refresh notification
