+++ libs/vscode/lsp-client/src/index.ts
@@ -1,2 +1,1 @@
 export * from './lib/nxls-client';
-export * from './lib/new-nxls-client';
